feat(backend): added functionality to CategoryService and CategoryController

+++ backend/src/main/java/at/qe/timeguess/services/CategoryService.java
@@ -1,15 +1,27 @@
 package at.qe.timeguess.services;
 
 import at.qe.timeguess.model.Category;
+import at.qe.timeguess.model.CompletedGame;
 import at.qe.timeguess.repositories.CategoryRepository;
+import at.qe.timeguess.repositories.CompletedGameRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Scope;
+import org.springframework.stereotype.Service;
 
 import java.util.Collection;
+import java.util.List;
 
+@Service
+@Scope("application")
 public class CategoryService {
+
+    // TODO: add PreAuthorize-annotations where needed
 
     @Autowired
     private CategoryRepository categoryRepository;
+
+    @Autowired
+    private CompletedGameRepository completedGameRepository;
 
     public Collection<Category> getAllCategories() {
         return this.categoryRepository.findAll();
@@ -23,11 +35,40 @@ public class CategoryService {
         return this.categoryRepository.findFirstByCategory(category);
     }
 
-    public Category saveCategory(final Category category) {
+    public Category saveCategory(final Category category) throws CategoryAlreadyExistsException{
+
+        if(this.getCategoryByName(category.getCategory()) != null) {
+            throw new CategoryAlreadyExistsException("This Category already exists!");
+        }
         return this.categoryRepository.save(category);
     }
 
-    public void deleteCategory(final Category category) {
+    public void deleteCategory(final Category category) throws CategoryAlreadyExistsException {
+
+        Collection<CompletedGame> allCompletedGames = this.completedGameRepository.findAll();
+
+        if(allCompletedGames.stream().anyMatch(completedGame -> completedGame.getCategory().getCategory().equals(category.getCategory()))) {
+            throw new CategoryAlreadyExistsException("This Category can not be deleted, because it is referenced in a completed game!");
+        }
         this.categoryRepository.delete(category);
     }
+
+    public class CategoryIsReferencedInCompletedGamesException extends Exception {
+
+        private static final long serialVersionUID = 1L;
+
+        public CategoryIsReferencedInCompletedGamesException(final String message) {
+            super(message);
+        }
+    }
+
+    public class CategoryAlreadyExistsException extends Exception {
+
+        private static final long serialVersionUID = 1L;
+
+        public CategoryAlreadyExistsException(final String message) {
+            super(message);
+        }
+    }
 }
+
